chore: Refactoring for better organization
